fix(inventory): make read-side Dapper queries run against PostgreSQL
Three latent defects surfaced by the new tenant-isolation integration
test (all in the read-side StockRead slice), none previously exercised
against a live engine:

- Dapper 2.1.x does not bind DateOnly/TimeOnly natively, so every query
  passing a DateOnly parameter (expiry @Today, consumption @From/@To)
  threw NotSupportedException. Register DateOnly/TimeOnly type handlers
  once at composition (AddSislabInfrastructure).
- PostgreSQL COUNT returns bigint; the read models expose the counts as
  int, so Dapper could not match the record constructors. Cast every
  projected COUNT (KPIs, per-location counts, COUNT(*) OVER() totals,
  movement counts) to ::int.
- Alias the paginated total_rows column as totalrows so Dapper binds it
  to the TotalRows record parameter (underscore names are not matched).
- The consumption report ran two statements in one round-trip sharing a
  WITH consumption CTE, but a CTE is scoped to a single statement, so
  the second SELECT failed with relation consumption does not exist.
  Prepend the CTE to both statements.

+++ src/Modules/Inventory/SISLAB.Modules.Inventory.Application/StockRead/GetLocationsSummaryQuery.cs
@@ -66,13 +66,13 @@ internal sealed class GetLocationsSummaryQueryHandler
             sl.name                                      AS name,
             sl.type                                      AS type,
             sl.is_active                                 AS isactive,
-            COUNT(si.id)                                 AS itemcount,
+            COUNT(si.id)::int                            AS itemcount,
             COUNT(si.id) FILTER (
                 WHERE si.expiry_year IS NOT NULL
                   AND si.expiry_month IS NOT NULL
                   AND @Today > (make_date(si.expiry_year, si.expiry_month, 1)
                                 + INTERVAL '1 month' - INTERVAL '1 day')::date
-            )                                            AS expireditemcount,
+            )::int                                       AS expireditemcount,
             (sl.type = @ControlledType)                  AS iscritical
         FROM inventory.storage_locations AS sl
         LEFT JOIN inventory.stock_items AS si
